+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### New features
 
 * Make the content of the information box configurable
+* Show the time left when rendering the video
 * Delete cached map tiles in preferences dialog
 * New configuration option to specify the user agent when accessing the TMS servers